- Music now fades in/out on a world change. It is still a bit buggy, though. At times one of the music tracks stops (I think it can be reproduced by changing twice in a short while: the transition back works and the volume traces back to its previous value correctly, but the next world change result is messy).
- Fixed (at last :D) that nasty crash that happened when destroying the Physics Subsystem's scene (specifically, the Ogre3DRenderSystem instance). 
It was caused by calls to PhysicsComponentCharacter::destroy(), where the destroyController() call deleted the nxOgreController instance: it turns out that in that call the attribute 'mRenderable' doesn't get destroyed in return, so it ends up as an orphan pointer. 

Because of this, when the OGRE3DPointRenderable destructor is called, the list of pointRenderables contain such 'orphan' pointers whose content is now corrupt, which leads to that annoying, seemingly random exception.

The solution at the moment (there are more general and elegant alternatives, which involve tweaking NxOgre yet again, but the proposed one is valid as well given our project's necessities)